Rename MarkCommentAsSeenMutation -> MarkCommentsAsSeenMutation
