+++ src/core/client/stream/tabs/Comments/Comment/CommentContainer.tsx
@@ -64,7 +64,7 @@ import EditCommentFormContainer from "./EditCommentForm";
 import FeaturedTag from "./FeaturedTag";
 import { isReplyFlattened } from "./flattenReplies";
 import IndentedComment from "./IndentedComment";
-import MarkCommentAsSeenMutation from "./MarkCommentAsSeenMutation";
+import MarkCommentAsSeenMutation from "./MarkCommentsAsSeenMutation";
 import MediaSectionContainer from "./MediaSection/MediaSectionContainer";
 import CaretContainer, {
   ModerationRejectedTombstoneContainer,
